p4i integration Integrate 259625 from 4.3 to main: Fixes:    removed compiler warning (mostly VS2005 and Win64 warnings) RevBy:    prasanth
[git-p4: depot-paths = "//depot/qt/main/": change = 259628]

+++ src/qt3support/text/q3textstream.cpp
@@ -306,9 +306,9 @@ qint64 QStringBuffer::readData( char *p, qint64 len )
 	return qint64(-1);
     }
 #endif
-    if ( pos() + len > s->length()*sizeof(QChar) ) {
+    if ( pos() + len > qint64(s->length()*sizeof(QChar)) ) {
 	// overflow
-	if ( pos() >= s->length()*sizeof(QChar) ) {
+	if ( pos() >= qint64(s->length()*sizeof(QChar)) ) {
 	    return -1;
 	} else {
 	    len = s->length()*2 - pos();
